refactor: type files' names updated
